feat: create hook useRoom
