+++ src/hooks/useRoom.ts
@@ -1,0 +1,51 @@
+import { useEffect, useState } from "react";
+import { database } from "../services/firebase";
+
+type FirebaseQuestions = Record<string, {
+  author: {
+    name: string,
+    avatar: string,
+  }
+  content: string;
+  isHighLighted: boolean,
+  isAnswered: boolean,
+}>
+
+type QuestionsType = {
+  id: string;
+  author: {
+    name: string,
+    avatar: string,
+  }
+  content: string;
+  isHighLighted: boolean,
+  isAnswered: boolean,
+}
+
+export function useRoom(roomId: string) {
+  const [questions, setQuestions] = useState<QuestionsType[]>([]);
+  const [title, setTitle] = useState('');
+
+  useEffect(() => {
+    const roomRef = database.ref(`rooms/${roomId}`);
+
+    roomRef.on('value', room => {
+      const databaseRoom = room.val();
+      const firebaseQuestions: FirebaseQuestions = databaseRoom.questions ?? {};
+      const parsedQuestions = Object.entries(firebaseQuestions)
+        .map(([key, value]) => {
+          return { 
+            id: key,
+            content: value.content,
+            author: value.author,
+            isAnswered: value.isAnswered,
+            isHighLighted: value.isHighLighted,
+          }
+        });
+        setTitle(databaseRoom.title);
+        setQuestions(parsedQuestions);
+    });
+  }, [roomId])
+
+  return { questions, title };
+}
