+++ src/pages/Room.tsx
@@ -7,27 +7,7 @@ import { useAuth } from '../hooks/useAuth';
 import { database } from '../services/firebase';
 import '../styles/rooms.scss';
 import {Question} from '../components/Question';
-
-type FirebaseQuestions = Record<string, {
-  author: {
-    name: string,
-    avatar: string,
-  }
-  content: string;
-  isHighLighted: boolean,
-  isAnswered: boolean,
-}>
-
-type QuestionsType = {
-  id: string;
-  author: {
-    name: string,
-    avatar: string,
-  }
-  content: string;
-  isHighLighted: boolean,
-  isAnswered: boolean,
-}
+import { useRoom } from '../hooks/useRoom';
 
 type RoomParams = {
   id: string,
@@ -37,32 +17,9 @@ export function Room() {
   const {user} = useAuth();
   const params = useParams<RoomParams>();
   const [newQuestion, setNewQuestion] = useState('');
-  const [questions, setQuestions] = useState<QuestionsType[]>([]);
-  const [title, setTitle] = useState('');
-
   const roomId = params.id;
+  const { title, questions } = useRoom(roomId);
 
-  useEffect(() => {
-    const roomRef = database.ref(`rooms/${roomId}`);
-
-    roomRef.once('value', room => {
-      const databaseRoom = room.val();
-      const firebaseQuestions: FirebaseQuestions = databaseRoom.questions ?? {};
-      const parsedQuestions = Object.entries(firebaseQuestions)
-        .map(([key, value]) => {
-          return { 
-            id: key,
-            content: value.content,
-            author: value.author,
-            isAnswered: value.isAnswered,
-            isHighLighted: value.isHighLighted,
-          }
-        });
-        setTitle(databaseRoom.title);
-        setQuestions(parsedQuestions);
-    });
-  }, [roomId, newQuestion])
-  
   async function handleSendQuestion(event: FormEvent) {
     event.preventDefault();
     if (newQuestion.trim() === '') {
